Add check if hash already exists in database

+++ foxogram-common/src/main/java/su/foxogram/services/StorageService.java
@@ -8,12 +8,20 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
+import su.foxogram.models.Channel;
+import su.foxogram.models.Message;
+import su.foxogram.models.User;
+import su.foxogram.repositories.ChannelRepository;
+import su.foxogram.repositories.MessageRepository;
+import su.foxogram.repositories.UserRepository;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
@@ -29,21 +37,36 @@ public class StorageService {
 
 	private final MinioAsyncClient minioClient;
 
+	private final UserRepository userRepository;
+
+	private final ChannelRepository channelRepository;
+
+	private final MessageRepository messageRepository;
+
 	@Autowired
-	public StorageService(MinioAsyncClient minioClient) {
+	public StorageService(MinioAsyncClient minioClient, UserRepository userRepository, ChannelRepository channelRepository, MessageRepository messageRepository) {
 		this.minioClient = minioClient;
+		this.userRepository = userRepository;
+		this.channelRepository = channelRepository;
+		this.messageRepository = messageRepository;
 	}
 
-	public String uploadFile(MultipartFile file, String bucketName) throws RuntimeException, IOException, ExecutionException, InterruptedException {
+	public String uploadFile(MultipartFile file, String bucketName) throws RuntimeException, IOException, ExecutionException, InterruptedException, NoSuchAlgorithmException {
 		byte[] byteArray = file.getBytes();
 		long startTime = System.currentTimeMillis();
+		String fileHash = getHash(byteArray);
 		log.info("Uploading file ({}) to bucket ({})", file.getOriginalFilename(), bucketName);
 
-		if (!isBucketExist(bucketName).get()) createBucket(bucketName);
+		if (isHashExists(fileHash)) {
+			log.info("Duplicate image ({}) found. Skipping upload...", fileHash);
+			return fileHash;
+		}
+
+		if (!isBucketExists(bucketName).get()) createBucket(bucketName);
 
 		try (InputStream pngStream = new ByteArrayInputStream(byteArray);
 			 InputStream webpStream = new ByteArrayInputStream(byteArray)) {
-			String fileHash = getHash(byteArray);
+
 
 			minioClient.putObject(
 					PutObjectArgs.builder().bucket(bucketName).object(fileHash + ".png").stream(
@@ -67,6 +90,14 @@ public class StorageService {
 		}
 	}
 
+	private boolean isHashExists(String hash) {
+		User user = userRepository.findByAvatar(hash);
+		Channel channel = channelRepository.findByIcon(hash);
+		List<String> attachments = getAllAttachments();
+
+		return user != null || channel != null || attachments.contains(hash);
+	}
+
 	private void createBucket(String bucketName) {
 		try {
 			minioClient.makeBucket(MakeBucketArgs.builder().bucket(bucketName).build());
@@ -75,7 +106,7 @@ public class StorageService {
 		}
 	}
 
-	private CompletableFuture<Boolean> isBucketExist(String bucketName) throws RuntimeException {
+	private CompletableFuture<Boolean> isBucketExists(String bucketName) throws RuntimeException {
 		try {
 			return minioClient.bucketExists(BucketExistsArgs.builder().bucket(bucketName).build());
 		} catch (Exception e) {
@@ -93,5 +124,17 @@ public class StorageService {
 		}
 
 		return hexString.toString();
+	}
+
+	private List<String> getAllAttachments() {
+		List<Message> messages = messageRepository.findAll();
+		List<String> attachments = new ArrayList<>();
+		for (Message message : messages) {
+			if (message.getAttachments() != null) {
+				attachments.addAll(message.getAttachments());
+			}
+		}
+
+		return attachments;
 	}
 }
